User commit

+++ src/main/java/com/example/demo/repository/ValidationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.entity.ValidationEntity;
 
 @Repository
-public interface ValidationRepository extends JpaRepository<ValidationEntity,Integer>{
+public interface ValidationRepository extends JpaRepository<ValidationEntity,Long>{
 
 }
